Create Hill Climb Fall Game
This game has 2 players. In this game there is a hill which consists of 10 steps and the players need to climb the hill until they reach the 10th step by rolling a dice to win. The twist here is, if any one of the player cross the 10th step then they fall into starting stage and start from 0.